finish user type controller
